feat:移动YakitRoute枚举到newRouteConstants,修改引用文件位置
