+++ app/renderer/src/main/src/components/layout/FuncDomain.tsx
@@ -59,7 +59,7 @@ import {showYakitModal} from "../yakitUI/YakitModal/YakitModalConfirm"
 import {MacKeyborad, WinKeyborad} from "../yakitUI/YakitEditor/editorUtils"
 import {ScrecorderModal} from "@/pages/screenRecorder/ScrecorderModal"
 import {useScreenRecorder} from "@/store/screenRecorder"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {RouteToPageProps} from "@/pages/layout/publicMenu/PublicMenu"
 import {useRunNodeStore} from "@/store/runNode"
 import emiter from "@/utils/eventBus/eventBus"
