+++ app/renderer/src/main/src/components/yakitUI/YakitEditor/YakitEditor.tsx
@@ -67,7 +67,7 @@ import {setEditorContext} from "@/utils/monacoSpec/yakEditor"
 import {YakParamProps} from "@/pages/plugins/pluginsType"
 import {usePageInfo} from "@/store/pageInfo"
 import {shallow} from "zustand/shallow"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {HighLightText} from "@/components/HTTPFlowDetail"
 import {useStore} from "@/store/editorState"
 import {CloudDownloadIcon} from "@/assets/newIcon"
