+++ app/renderer/src/main/src/pages/MainOperator.tsx
@@ -37,7 +37,7 @@ import {ControlOperation} from "@/pages/dynamicControl/DynamicControl"
 import {YakitHintModal} from "@/components/yakitUI/YakitHint/YakitHintModal"
 import {useScreenRecorder} from "@/store/screenRecorder"
 import PublicMenu, {RouteToPageProps} from "./layout/publicMenu/PublicMenu"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {YakChatCS} from "@/components/yakChat/chatCS"
 import yakitCattle from "../assets/yakitCattle.png"
 import {MainOperatorContent} from "./layout/mainOperatorContent/MainOperatorContent"
