+++ app/renderer/src/main/src/pages/assetViewer/DomainAssetPage.tsx
@@ -12,7 +12,7 @@ import {onRemoveToolFC} from "../../utils/deleteTool"
 
 import styles from "./DomainAssetPage.module.scss"
 import emiter from "@/utils/eventBus/eventBus"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 
 export interface Domain {
     ID?: number
