+++ app/renderer/src/main/src/pages/fuzzer/FuzzerSequence/FuzzerPageSetting.tsx
@@ -1,4 +1,4 @@
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {PageNodeItemProps, usePageInfo} from "@/store/pageInfo"
 import {useDebounceFn, useInViewport, useMemoizedFn, useUpdateEffect} from "ahooks"
 import React, {useEffect, useRef, useState} from "react"
