+++ app/renderer/src/main/src/pages/fuzzer/components/ShareImport/index.tsx
@@ -9,7 +9,7 @@ import {useStore} from "@/store"
 import {YakitInput} from "@/components/yakitUI/YakitInput/YakitInput"
 import {YakitButton} from "@/components/yakitUI/YakitButton/YakitButton"
 import {showYakitModal} from "@/components/yakitUI/YakitModal/YakitModalConfirm"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 
 const layout = {
     labelCol: {span: 5},
