+++ app/renderer/src/main/src/pages/fuzzer/components/ShareImportExportData/index.tsx
@@ -20,7 +20,7 @@ import {isCommunityEdition} from "@/utils/envfile"
 import {saveABSFileAnotherOpen} from "@/utils/openWebsite"
 import {Uint8ArrayToString, StringToUint8Array} from "@/utils/str"
 import {NewHTTPPacketEditor} from "@/utils/editors"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {FuzzerRequestProps} from "../../HTTPFuzzerPage"
 import {AdvancedConfigValueProps} from "../../HttpQueryAdvancedConfig/HttpQueryAdvancedConfigType"
 import emiter from "@/utils/eventBus/eventBus"
