+++ app/renderer/src/main/src/pages/invoker/batch/BatchExecuteByFilter.tsx
@@ -24,7 +24,7 @@ import {YakitLogFormatter} from "../YakitLogFormatter";
 import "./BatchExecuteByFilter.css"
 import {Risk} from "../../risks/schema";
 import {RisksViewer} from "@/pages/risks/RisksViewer";
-import { YakitRoute, YakitRouteToPageInfo } from "@/routes/newRoute";
+import {YakitRoute} from "@/routes/newRouteConstants";
 
 const {ipcRenderer} = window.require("electron");
