+++ app/renderer/src/main/src/pages/layout/mainOperatorContent/MainOperatorContent.tsx
@@ -21,7 +21,6 @@ import {
 import styles from "./MainOperatorContent.module.scss"
 import {
     YakitRouteToPageInfo,
-    YakitRoute,
     SingletonPageRoute,
     NoPaddingRoute,
     ComponentParams,
@@ -92,7 +91,8 @@ import {defaultBrutePageInfo} from "@/defaultConstants/NewBrute"
 import {defaultScanPortPageInfo} from "@/defaultConstants/NewPortScan"
 import {defaultPocPageInfo} from "@/defaultConstants/YakPoC"
 import {defaultSpaceEnginePageInfo} from "@/defaultConstants/SpaceEnginePage"
-import { defaultSimpleDetectPageInfo } from "@/defaultConstants/SimpleDetect"
+import {defaultSimpleDetectPageInfo} from "@/defaultConstants/SimpleDetectConstants"
+import {YakitRoute} from "@/routes/newRouteConstants"
 
 const TabRenameModalContent = React.lazy(() => import("./TabRenameModalContent"))
 const PageItem = React.lazy(() => import("./renderSubPage/RenderSubPage"))
