+++ app/renderer/src/main/src/pages/layout/publicMenu/ExtraMenu.tsx
@@ -2,7 +2,7 @@ import React, {useMemo, useState} from "react"
 import {YakitPopover} from "@/components/yakitUI/YakitPopover/YakitPopover"
 import {YakitButton} from "@/components/yakitUI/YakitButton/YakitButton"
 import {YakitMenu} from "@/components/yakitUI/YakitMenu/YakitMenu"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {onImportShare} from "@/pages/fuzzer/components/ShareImport"
 import {useMemoizedFn} from "ahooks"
 import {RouteToPageProps} from "./PublicMenu"
