+++ app/renderer/src/main/src/pages/layout/publicMenu/MenuPlugin.tsx
@@ -5,7 +5,7 @@ import {YakitPopover} from "@/components/yakitUI/YakitPopover/YakitPopover"
 import {useMemoizedFn} from "ahooks"
 import {EnhancedPublicRouteMenuProps} from "./utils"
 import {Avatar, Tooltip} from "antd"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {RouteToPageProps} from "./PublicMenu"
 import {LoadingOutlined} from "@ant-design/icons"
 import {CodeGV, RemoteGV} from "@/yakitGV"
