+++ app/renderer/src/main/src/pages/plugins/editDetails/PluginEditDetails.tsx
@@ -45,7 +45,7 @@ import {PluginModifyInfo, PluginModifySetting} from "../baseTemplate"
 import emiter from "@/utils/eventBus/eventBus"
 import {toolDelInvalidKV} from "@/utils/tool"
 import {useSubscribeClose} from "@/store/tabSubscribe"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {DefaultTypeList, GetPluginLanguage, PluginGV, pluginTypeToName} from "../builtInData"
 import {PageNodeItemProps, usePageInfo} from "@/store/pageInfo"
 import {shallow} from "zustand/shallow"
