+++ app/renderer/src/main/src/pages/plugins/local/PluginsLocal.tsx
@@ -61,7 +61,7 @@ import {YakitCheckbox} from "@/components/yakitUI/YakitCheckbox/YakitCheckbox"
 import {showYakitModal} from "@/components/yakitUI/YakitModal/YakitModalConfirm"
 import emiter from "@/utils/eventBus/eventBus"
 import {PluginLocalUpload, PluginLocalUploadSingle} from "./PluginLocalUpload"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {DefaultTypeList, PluginGV, defaultFilter, defaultSearch} from "../builtInData"
 import {RemoteGV} from "@/yakitGV"
 import {randomString} from "@/utils/randomUtil"
