+++ app/renderer/src/main/src/pages/plugins/local/PluginsLocalDetail.tsx
@@ -30,7 +30,7 @@ import {getRemoteValue} from "@/utils/kv"
 import {YakitSpin} from "@/components/yakitUI/YakitSpin/YakitSpin"
 import {ExclamationCircleOutlined} from "@ant-design/icons"
 import emiter from "@/utils/eventBus/eventBus"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {SolidCloudpluginIcon, SolidPrivatepluginIcon} from "@/assets/icon/colors"
 import PluginTabs from "@/components/businessUI/PluginTabs/PluginTabs"
 import {LocalPluginExecute} from "./LocalPluginExecute"
