+++ app/renderer/src/main/src/pages/plugins/online/PluginsOnline.tsx
@@ -65,7 +65,7 @@ import {TypeSelectOpt} from "../funcTemplateType"
 import {DefaultTypeList, PluginGV, defaultSearch, funcSearchType, pluginTypeToName} from "../builtInData"
 import {getRemoteValue, setRemoteValue} from "@/utils/kv"
 import emiter from "@/utils/eventBus/eventBus"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {YakitButton} from "@/components/yakitUI/YakitButton/YakitButton"
 import {isCommunityEdition} from "@/utils/envfile"
 import {YakitHint} from "@/components/yakitUI/YakitHint/YakitHint"
