+++ app/renderer/src/main/src/pages/plugins/online/PluginsOnlineDetail.tsx
@@ -14,7 +14,7 @@ import {thousandthConversion} from "../pluginReducer"
 import {useStore} from "@/store"
 import {LoadingOutlined} from "@ant-design/icons"
 import emiter from "@/utils/eventBus/eventBus"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {onlineUseToLocalDetail} from "../utils"
 import PluginTabs from "@/components/businessUI/PluginTabs/PluginTabs"
 import {PluginLog} from "../log/PluginLog"
