+++ app/renderer/src/main/src/pages/plugins/pluginBatchExecutor/pluginBatchExecutor.tsx
@@ -38,7 +38,7 @@ import {PluginExecuteResult} from "../operator/pluginExecuteResult/PluginExecute
 import {ExpandAndRetract, ExpandAndRetractExcessiveState} from "../operator/expandAndRetract/ExpandAndRetract"
 import {PageNodeItemProps, PluginBatchExecutorPageInfoProps, usePageInfo} from "@/store/pageInfo"
 import {shallow} from "zustand/shallow"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {StreamResult} from "@/hook/useHoldGRPCStream/useHoldGRPCStreamType"
 import {PluginLocalListDetails} from "../operator/PluginLocalListDetails/PluginLocalListDetails"
 import {PluginExecuteLog} from "@/pages/securityTool/yakPoC/YakPoC"
