+++ app/renderer/src/main/src/pages/plugins/user/PluginUserDetail.tsx
@@ -14,7 +14,7 @@ import cloneDeep from "lodash/cloneDeep"
 import {PluginUserDetailProps, UserBackInfoProps} from "./PluginUserType"
 import {useStore} from "@/store"
 import emiter from "@/utils/eventBus/eventBus"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {onlineUseToLocalDetail} from "../utils"
 import {LoadingOutlined} from "@ant-design/icons"
 import {SolidPrivatepluginIcon} from "@/assets/icon/colors"
