+++ app/renderer/src/main/src/pages/securityTool/newPortScan/NewPortScan.tsx
@@ -50,7 +50,7 @@ import {apiGetGlobalNetworkConfig} from "@/pages/spaceEngine/utils"
 import {GlobalNetworkConfig} from "@/components/configNetwork/ConfigNetworkPage"
 import {shallow} from "zustand/shallow"
 import {PageNodeItemProps, ScanPortPageInfoProps, usePageInfo} from "@/store/pageInfo"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 import {pluginTypeFilterList} from "@/defaultConstants/PluginBatchExecutor"
 import {defaultScanPortPageInfo} from "@/defaultConstants/NewPortScan"
