+++ app/renderer/src/main/src/pages/yakitStore/viewers/WebsiteTree.tsx
@@ -14,7 +14,7 @@ import {ExportExcel} from "../../../components/DataExport/DataExport"
 import {ChevronDownIcon} from "@/assets/newIcon"
 import "./WebsiteTreeStyle.css"
 import emiter from "@/utils/eventBus/eventBus"
-import {YakitRoute} from "@/routes/newRoute"
+import {YakitRoute} from "@/routes/newRouteConstants"
 
 export interface WebsiteTreeViewerProp {
     pageMode?: boolean
